add break row modes and start of some documentation
