+++ kernel/kernel.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include "../libc/itoa.c"
 #include "../libc/memset.c"
-
+#include "../libc/strlen.c"
 
 /* Check if the compiler thinks you are targeting the wrong operating system. */
 #if defined(__linux__)
@@ -33,155 +33,8 @@ typedef s32 				b32;
 #define local_persist 	static
 #define global_variable static
 
-/* Hardware text mode color constants. */
-enum vga_color {
-	VGA_COLOR_BLACK = 0,
-	VGA_COLOR_BLUE = 1,
-	VGA_COLOR_GREEN = 2,
-	VGA_COLOR_CYAN = 3,
-	VGA_COLOR_RED = 4,
-	VGA_COLOR_MAGENTA = 5,
-	VGA_COLOR_BROWN = 6,
-	VGA_COLOR_LIGHT_GREY = 7,
-	VGA_COLOR_DARK_GREY = 8,
-	VGA_COLOR_LIGHT_BLUE = 9,
-	VGA_COLOR_LIGHT_GREEN = 10,
-	VGA_COLOR_LIGHT_CYAN = 11,
-	VGA_COLOR_LIGHT_RED = 12,
-	VGA_COLOR_LIGHT_MAGENTA = 13,
-	VGA_COLOR_LIGHT_BROWN = 14,
-	VGA_COLOR_WHITE = 15,
-};
-
-static inline uint8_t vga_entry_color(enum vga_color fg, enum vga_color bg)
-{
-	return fg | bg << 4;
-}
-
-static inline uint16_t vga_entry(unsigned char uc, uint8_t color)
-{
-	return (uint16_t) uc | (uint16_t) color << 8;
-}
-
-size_t strlen(const char* str)
-{
-	size_t len = 0;
-	while (str[len])
-		len++;
-	return len;
-}
-
-static const size_t VGA_WIDTH = 80;
-static const size_t VGA_HEIGHT = 25;
-
-size_t terminal_row;
-size_t terminal_column;
-uint8_t terminal_color;
-uint16_t* terminal_buffer;
-
-void terminal_initialize(void)
-{
-	terminal_row = 0;
-	terminal_column = 0;
-	terminal_color = vga_entry_color(VGA_COLOR_LIGHT_GREY, VGA_COLOR_BLACK);
-	terminal_buffer = (uint16_t*) 0xB8000;
-	for (size_t y = 0; y < VGA_HEIGHT; y++) {
-		for (size_t x = 0; x < VGA_WIDTH; x++) {
-			const size_t index = y * VGA_WIDTH + x;
-			terminal_buffer[index] = vga_entry(' ', terminal_color);
-		}
-	}
-}
-
-void terminal_setcolor(uint8_t color)
-{
-	terminal_color = color;
-}
-
-void terminal_putentryat(char c, uint8_t color, size_t x, size_t y)
-{
-	const size_t index = y * VGA_WIDTH + x;
-	terminal_buffer[index] = vga_entry(c, color);
-}
-
-void scroll_down()
-{
-	for (size_t y = 1; y < VGA_HEIGHT; ++y)
-	for (size_t x = 0; x < VGA_WIDTH; ++x)
-	{
-		const size_t up_one_row_index = (y-1) * VGA_WIDTH + x;
-		const size_t index = y * VGA_WIDTH + x;
-
-		terminal_buffer[up_one_row_index] = terminal_buffer[index];
-	}
-
-	terminal_row -= 1;
-}
-
-void terminal_putchar(char c)
-{
-	terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
-	if (++terminal_column == VGA_WIDTH) {
-		terminal_column = 0;
-		if (++terminal_row == VGA_HEIGHT)
-			terminal_row += 1;
-	}
-}
-
-void terminal_write(const char* data, size_t size)
-{
-	for (size_t i = 0; i < size; i++)
-	{
-		switch (data[i])
-		{
-			case '\n':
-			{
-				terminal_row += 1;
-				terminal_column = 0;
-
-				break;
-			}
-			case '\t':
-			{
-				terminal_column += 4;
-
-				if (terminal_column >= VGA_WIDTH)
-				{
-					terminal_row += 1;
-					terminal_column = 0;
-				}
-
-				break;
-			}
-			default:
-			{
-				terminal_putchar(data[i]);
-				break;
-			}
-		}
-
-		if (terminal_row + 1 >= VGA_HEIGHT)
-		{
-			scroll_down();
-		}
-	}
-}
-
-void terminal_writestring(const char* data)
-{
-	terminal_write(data, strlen(data));
-}
-
-void terminal_writenumber(const size_t number)
-{
-  	char buffer [33];
-
-  	const char * c = itoa(number, buffer, 10);
-
-	terminal_writestring(c);
-}
 #include "idt.c"
-
+#include "terminal.c"
 #include "interrupt.c"
 
 void kernel_main(void) 
@@ -191,14 +44,40 @@ void kernel_main(void)
 	
 	idt_initialize();
 
-	/* Newline support is left as an exercise. */
 	terminal_writestring("Welcome to Developer OS.\n");
-	terminal_writestring("This is a super long line that should automatically break to a new row when it hits the end of the terminal. If it doesn't something is wrong!\n");
-	terminal_writestring("This\n should\n handle\n the\n newline\n character.\n");
+
+	terminal_writestring("\n\n");
+
+	terminal_writestring("This output is testing so that the terminal escape");
+	terminal_writestring("This\nshould\nhandle\nthe\nnewline\ncharacter.\n");
   
 	terminal_writestring("Tabbing\t like\t crazy.\t PS.\t switch\t is\t the\t best\t control\t flow\t operator\t in\t the\t C\t language\t.\n");
+
+	terminal_writestring("A\nB\nC\nD\nE\nF\nG\nH\nI\nJ\nK\nL\nM\nN\nO\nP\nQ\nR\nS\nT\nU\nV\nW\nX\nY\nZ");
 	
+	terminal_writestring("\n\n");
+
+	terminal_writestring("This is an extra super long line and it should automatically break to a new row when it hits the end of the terminal row. If it doesn't, something is wrong! Since we're using the \"break row on word\" setting, these sentences shouldn't be broken in the middle of a word, but rather before / after a whole word. As a matter of fact, I can write a whole essay here, and it will still work as expected! Isn't that just awesome!\n\nI can aslo jump down a few lines. Look how this word does not get broken up: Omphaloskepsis");
+
+	terminal_writestring("\n\n");
+
+	terminal_writestring("Here is a number: ");
+
 	terminal_writenumber(11230);
-	terminal_writestring("\n");
+
+	terminal_writestring("\n\n");
+
+	terminal_writestring("This works when seperated into another file as well! Yay. unity build!");
+
+	terminal_writestring("\n\n");
+
+	terminal_setcolor(VGA_COLOR_CYAN);
+
+	terminal_writestring("Here is a message in the color cyan.");
+
+	terminal_writestring("\n\n");
+
+	terminal_setcolor(VGA_COLOR_BLACK);
+
 	asm volatile ("int $0x3"); // NOTE(Oskar): Triggers an ISR interrupt.
 }
