+++ libc/strlen.c
@@ -1,0 +1,10 @@
+size_t strlen(const char* str)
+{
+	size_t len = 0;
+	while (str[len])
+	{
+		len++;
+	}
+	
+	return len;
+}
